+++ module/pdf-refs.js
@@ -1,4 +1,5 @@
 import * as Settings from '../lib/miscellaneous-settings.js'
+import { PDFViewerSheet } from './pdf/sheet.js'
 
 export const SJGProductMappings = {
   ACT1: 'http://www.warehouse23.com/products/gurps-action-1-heroes',
@@ -71,11 +72,11 @@ export function handleOnPdf(event) {
  * @param {string} links
  */
 export function handlePdf(links) {
-  // @ts-ignore
-  if (!ui.PDFoundry) {
-    ui.notifications?.warn('PDFoundry must be installed and configured to use links.')
-    return
-  }
+  // // @ts-ignore
+  // if (!ui.PDFoundry) {
+  // 	ui.notifications?.warn('PDFoundry must be installed and configured to use links.')
+  // 	return
+  // }
 
   // Just in case we get sent multiple links separated by commas, we will open them all
   links.split(',').forEach(link => {
@@ -104,13 +105,22 @@ export function handlePdf(links) {
         page += 2
       } else page -= 335
     }
-    // @ts-ignore
-    const pdf = ui.PDFoundry.findPDFDataByCode(book)
-    if (pdf === undefined) {
+    const pdfPages = []
+    game.journal.forEach(j => {
+      j.pages.forEach(p => {
+        if (p.type === 'pdf') pdfPages.push(p)
+      })
+    })
+    let journalPage = null
+    if (pdfPages.length) journalPage = pdfPages.find(e => e.system.code === book)
+    if (journalPage) {
+      const viewer = new PDFViewerSheet(journalPage, { pageNumber: page })
+      viewer.render(true)
+    } else {
       let url = GURPS.SJGProductMappings[book]
-      if (!url) url = 'http://www.warehouse23.com/products?taxons%5B%5D=558398545-sb' // The main GURPS page
-      window.open(url, '_blank')
-      // @ts-ignore
-    } else ui.PDFoundry.openPDF(pdf, { page })
+      if (url)
+        // url = 'http://www.warehouse23.com/products?taxons%5B%5D=558398545-sb' // The main GURPS page
+        window.open(url, '_blank')
+    }
   })
 }